More sounds, improvements to basement door
Basement door is now it's own actor. Fixed problem where if the player was standing on top of the basement door while finishing the puzzle, the basement door would push them up. There are now trigger volumes to detect if the player is a safe distance away from the door before allowing the door to open. A sound plays if the player finishes the puzzle, but is standing on top of the door, disabling it from opening.

Added whisper to basement and looping music that plays after picking up the entrance/exit key.